added tests

+++ code/battle-manager/src/main/scripts/destroyer.sh
@@ -2,6 +2,11 @@
 
 # $1 -> directory where the zip
 
+if [ $# -ne 1 ]; then
+    echo "[ERROR] Usage: $0 <directory>"
+    exit 1
+fi
+
 cd || exit 1
 
 rm -rf "$1"
